backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
bcrypt==4.1.2
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
stripe==7.8.0
gunicorn==21.2.0
structlog==23.2.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
cryptography==46.0.5
